Fill-in combined address names in specs when needed
In solidusio/solidus#3524 we introduced a new preference that allows
to use the combined version of firstname/lastname.
This commit updates checkout feature specs to reflect this preference.
Once the split version will be deprecated and removed, we can
probably also revert this commit and use the combined version only.